fix(examples): rename FileTree example file
