+++ packages/examples/src/file-tree-example.vue
