+++ packages/examples/src/index.ts
@@ -17,9 +17,9 @@ export { default as ClaudeExample } from './demo-claude.vue'
 export { default as GrokExample } from './demo-grok.vue'
 export { default as EnvironmentVariables } from './environment-variables.vue'
 export { default as FileTreeBasic } from './file-tree-basic.vue'
+export { default as FileTreeExample } from './file-tree-example.vue'
 export { default as FileTreeExpanded } from './file-tree-expanded.vue'
 export { default as FileTreeSelection } from './file-tree-selection.vue'
-export { default as FileTreeExample } from './file-tree.vue'
 export { default as Image } from './image.vue'
 export { default as InlineCitation } from './inline-citation.vue'
 export { default as LoaderCustomStyling } from './loader-custom-styling.vue'
